Feat: BlockListener Listens All events and syncing DB works good

+++ server/src/migrations/20230118144920-Swap.js
@@ -37,7 +37,7 @@ module.exports = {
         allowNull: false,
       },
       dropRate: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10, 4),
         allowNull: false,
       },
       premiumInterval: {
